Abstracao das classes
